Throw an error if we fail to access Theme API
